backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
motor==3.7.1
pymongo==4.12.1
pydantic[email]==2.11.3
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-dotenv==1.1.0
python-multipart==0.0.20
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
gunicorn==23.0.0
